refactor code out of main.cpp
