+++ src/blockrelay/graphene.cpp
@@ -21,6 +21,7 @@
 #include "txorphanpool.h"
 #include "util.h"
 #include "utiltime.h"
+#include "validation/validation.h"
 
 static bool ReconstructBlock(CNode *pfrom, const bool fXVal, int &missingCount, int &unnecessaryCount);
