+++ src/test/checkblock_tests.cpp
@@ -9,6 +9,7 @@
 #include "primitives/block.h"
 #include "test/test_bitcoin.h"
 #include "utiltime.h"
+#include "validation/validation.h"
 
 #include <cstdio>
