+++ src/test/exploit_tests.cpp
@@ -20,6 +20,7 @@
 #include "unlimited.h"
 #include "util.h"
 #include "utilstrencodings.h"
+#include "validation/validation.h"
 #include "version.h"
 
 #include "test/test_bitcoin.h"
