+++ src/test/miner_tests.cpp
@@ -17,6 +17,7 @@
 #include "uint256.h"
 #include "util.h"
 #include "utilstrencodings.h"
+#include "validation/validation.h"
 
 #include "test/test_bitcoin.h"
