+++ src/txadmission.cpp
@@ -12,6 +12,7 @@
 #include "init.h"
 #include "main.h" // for cs_main
 #include "net.h"
+#include "parallel.h"
 #include "requestManager.h"
 #include "respend/respenddetector.h"
 #include "timedata.h"
@@ -19,6 +20,7 @@
 #include "unlimited.h"
 #include "util.h"
 #include "utiltime.h"
+#include "validation/validation.h"
 #include "validationinterface.h"
 #include <map>
 #include <string>
