+++ src/unlimited.cpp
@@ -41,6 +41,7 @@
 #include "util.h"
 #include "utilmoneystr.h"
 #include "utilstrencodings.h"
+#include "validation/validation.h"
 #include "validationinterface.h"
 #include "version.h"
